move friendship actions to dedicated module and update authentication error messages
