+++ containers/frontend/web/src/i18n/messages/ca.json
@@ -478,7 +478,8 @@
     "AUTH_NO_PENDING_RECOVER": "No s'ha trobat cap sol·licitud de restabliment. Torna a començar.",
     "AUTH_EMAIL_NOT_VERIFIED": "Verifica el teu correu abans d'iniciar sessió.",
     "AUTH_TOKEN_EXPIRED": "L'enllaç de verificació no és vàlid o ha caducat.",
-    "AUTH_FORBIDDEN": "L'enllaç de verificació no és vàlid o ha caducat.",
+    "AUTH_UNAUTHORIZED": "Has d'iniciar sessió per veure aquest contingut.",
+    "AUTH_FORBIDDEN": "No tens permís per accedir a aquest recurs.",
     "AUTH_INTERNAL_ERROR": "Alguna cosa ha anat malament. Torna-ho a provar.",
     "VALIDATION_ERROR": "Error de validació: revisa els camps i intenta de nou.",
     "FETCH_FAILED": "Ha fallat la sol·licitud a l'API. Torna-ho a provar més tard.",
