+++ containers/frontend/web/src/i18n/messages/en.d.json.ts
@@ -480,7 +480,8 @@ declare const messages: {
     "AUTH_NO_PENDING_RECOVER": "No password reset request was found. Please start again.",
     "AUTH_EMAIL_NOT_VERIFIED": "Verify your email before signing in.",
     "AUTH_TOKEN_EXPIRED": "The verification link is invalid or expired.",
-    "AUTH_FORBIDDEN": "The verification link is invalid or expired.",
+    "AUTH_UNAUTHORIZED": "You must be logged in to view this content.",
+    "AUTH_FORBIDDEN": "You do not have permission to access this resource.",
     "AUTH_INTERNAL_ERROR": "Something went wrong. Please try again.",
     "VALIDATION_ERROR": "Validation error: please check the fields and try again.",
     "FETCH_FAILED": "API request failed. Try again later.",
