+++ containers/frontend/web/src/i18n/messages/es.json
@@ -477,7 +477,8 @@
     "AUTH_NO_PENDING_RECOVER": "No se encontró ninguna solicitud de restablecimiento. Empieza de nuevo.",
     "AUTH_EMAIL_NOT_VERIFIED": "Verifica tu correo antes de iniciar sesión.",
     "AUTH_TOKEN_EXPIRED": "El enlace de verificación no es válido o ha caducado.",
-    "AUTH_FORBIDDEN": "El enlace de verificación no es válido o ha caducado.",
+    "AUTH_UNAUTHORIZED": "Debes iniciar sesión para ver este contenido.",
+    "AUTH_FORBIDDEN": "No tienes permiso para acceder a este recurso.",
     "AUTH_INTERNAL_ERROR": "Algo salió mal. Vuelve a intentarlo.",
     "VALIDATION_ERROR": "Error de validación: revisa los campos e inténtalo de nuevo.",
     "FETCH_FAILED": "Ha fallado la solicitud a la API. Inténtalo de nuevo más tarde.",
